requirements.txt
ultralytics=8.3.197
pyyaml